refactor : 로또 구매 금액의 타입 변경 - 로또 구매 금액의 타입을 Long형으로 변경

+++ src/main/java/lotto/Application.java
@@ -19,7 +19,7 @@ public class Application {
     }
 
     public static void purchaseLotto() {
-        Integer totalCost = lottoInputController.inputSingleValue(inputView::inputTotalCost);
+        Long totalCost = lottoInputController.inputSingleValue(inputView::inputTotalCost);
         List<Lotto> purchased = lottoMachineController.purchase(totalCost);
 
         lottoCheckerController.setPurchasedLotto(purchased, totalCost);
